Added hydra, bandit, beholder to initializer (untested)

+++ api/Ganymede.Api.Models/Monster/SpecialTraits/Spellcasting/InnateSpellcastingModel.cs
@@ -20,7 +20,7 @@ namespace Ganymede.Api.Models.Monster.SpecialTraits.Spellcasting
                 {
                     ID = s.ID,
                     Psionic = s.Psionic,
-                    SpellcastingAbility = s.SpellcastingAbility,
+                    SpellcastingAbility = (int)s.SpellcastingAbility,
                     SpellcastingType = (int)s.SpellcastingType
                 }));
             CreateMap<InnateSpellcasting, InnateSpellcastingModel>()
